requirements.txt
streamlit
pyannote.audio
openai-whisper
langdetect
matplotlib
python-docx
pandas
huggingface-hub
